+++ examples/fn_02/Makefile
@@ -1,7 +1,7 @@
 CC = gcc
-SRCS = ../../slave/src/mb_slave.c main.c
+SRCS = ../../slave/src/mb_slave.c ../../slave/src/mb_crc.c read_input_status_02.c
 INCS = ../../slave/inc/
-EXE = bin
+EXE = read_input_status_02
 
 all:
 	@ $(CC) -o $(EXE) $(SRCS) -I $(INCS)
